requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
python-dotenv==1.1.0
requests==2.32.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
MarkupSafe==3.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
click==8.2.1
blinker==1.9.0
Werkzeug==3.1.3
SQLAlchemy==2.0.41
greenlet==3.2.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
Flask-Migrate==4.1.0